remove token_optimizer
